+++ stytra/examples/combined_exp.py
@@ -1,6 +1,9 @@
 from stytra import Stytra, Protocol
-from stytra.stimulation.stimuli.visual import StimulusCombiner, MovingGratingStimulus, \
-    HighResMovingWindmillStimulus
+from stytra.stimulation.stimuli.visual import (
+    StimulusCombiner,
+    MovingGratingStimulus,
+    HighResMovingWindmillStimulus,
+)
 import pandas as pd
 import numpy as np
 
@@ -16,15 +19,12 @@ class CombinedProtocol(Protocol):
 
         df = pd.DataFrame(dict(t=t, vel_x=vel))
 
-        s_a = MovingGratingStimulus(
-                df_param=df,
-                clip_mask=[0, 0, 1, 0.5])
+        s_a = MovingGratingStimulus(df_param=df, clip_mask=[0, 0, 1, 0.5])
 
         df = pd.DataFrame(dict(t=t, vel_x=-vel))
         s_b = MovingGratingStimulus(
-            df_param=df,
-            grating_angle=45,
-            clip_mask=[0, 0.5, 1, 0.5])
+            df_param=df, grating_angle=45, clip_mask=[0, 0.5, 1, 0.5]
+        )
 
         p = 1
         d = 5