+++ stytra/examples/custom_exp.py
@@ -14,6 +14,11 @@ class FlashProtocol(Protocol):
         return [Stimulus(duration=5.0)]
 
 
+# Little demonstration on how to use a custom experiment to bypass standard
+# launching through Stytra class. This little experiment simply add an additional
+# message box warning the user to confirm before running the protocol.
+
+
 class CustomExperiment(Experiment):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -22,12 +27,12 @@ class CustomExperiment(Experiment):
 
     def start_protocol(self):
         self.start = False
-
+        # Show message box with PyQt:
         msgBox = QMessageBox()
-        msgBox.setText("Start the protocol when ready")
+        msgBox.setText("Start the protocol when ready!")
         msgBox.setStandardButtons(QMessageBox.Ok)
-        ret = msgBox.exec_()
-        super().start_protocol()
+        _ = msgBox.exec_()
+        super().start_protocol()  # call the super() start_protocol method
 
 
 if __name__ == "__main__":