+++ stytra/examples/display_camera_exp.py
@@ -11,8 +11,7 @@ class Nostim(Protocol):
     # parameters that control camera, tracking, monitor, etc.
     # In this particular case, we add a stream of frames from one example
     # movie saved in stytra assets.
-    stytra_config = dict(
-        camera=dict(type="spinnaker"))
+    stytra_config = dict(camera=dict(type="spinnaker"))
 
     #  For a streaming from real cameras connected to the computer, specify camera type, e.g.:
     # stytra_config = dict(camera=dict(type="ximea"))