+++ stytra/examples/zmq_trigger_exp.py
@@ -8,8 +8,8 @@ class FlashProtocol(Protocol):
 
     def __init__(self):
         super().__init__()
-        self.period_sec = 3.
-        self.flash_duration = 2.
+        self.period_sec = 3.0
+        self.flash_duration = 2.0
 
     def get_stim_sequence(self):
         stimuli = [