+++ stytra/stimulation/__init__.py
@@ -92,7 +92,6 @@ class ProtocolRunner(QObject):
         self.past_stimuli_elapsed = None  # time elapsed in previous stimuli
         self.dynamic_log = None  # dynamic log for stimuli
 
-        self._set_new_protocol(self.protocol)
         self.update_protocol()
         self.protocol.sig_param_changed.connect(self.update_protocol)
 
@@ -104,46 +103,23 @@ class ProtocolRunner(QObject):
         self.framerate_rec = FramerateRecorder()
         self.framerate_acc = FramerateAccumulator(experiment=self.experiment)
 
-    def _set_new_protocol(self, protocol):
-        """Set new Protocol.
-
-        Parameters
-        ----------
-        protocol : :obj:`stytra.experiment.Protocol`
-            Protocol to be set.
-
-        """
-        if protocol is not None:
-            self.protocol = protocol
-
-            # Connect changes to protocol parameters to update function:
-            self.protocol.sig_param_changed.connect(self.update_protocol)
-            self.experiment.dc.add(self.protocol)
-
-            # Why were we resetting here?
-            self.reset()
-            self.update_protocol()
-            self.sig_protocol_updated.emit()
-
     def update_protocol(self):
-        """Update current Protocol (get a new stimulus list if protocol
-        exist.
+        """Update current Protocol (get a new stimulus list)
         """
-        if self.protocol is not None:
-            self.stimuli = self.protocol._get_stimulus_list()
+        self.stimuli = self.protocol._get_stimulus_list()
 
-            self.current_stimulus = self.stimuli[0]
+        self.current_stimulus = self.stimuli[0]
 
-            # pass experiment to stimuli for calibrator and asset folders:
-            for stimulus in self.stimuli:
-                stimulus.initialise_external(self.experiment)
+        # pass experiment to stimuli for calibrator and asset folders:
+        for stimulus in self.stimuli:
+            stimulus.initialise_external(self.experiment)
 
-            if self.dynamic_log is None:
-                self.dynamic_log = DynamicLog(self.stimuli, experiment=self.experiment)
-            else:
-                self.dynamic_log.update_stimuli(self.stimuli)  # new stimulus log
+        if self.dynamic_log is None:
+            self.dynamic_log = DynamicLog(self.stimuli, experiment=self.experiment)
+        else:
+            self.dynamic_log.update_stimuli(self.stimuli)  # new stimulus log
 
-            self.sig_protocol_updated.emit()
+        self.sig_protocol_updated.emit()
 
     def reset(self):
         """Make the protocol ready to start again. Reset all ProtocolRunner
@@ -305,22 +281,6 @@ class ProtocolRunner(QObject):
             duration += stim.duration
         return duration
 
-
-    #def get_duration(self):
-    #    """Get total duration of the protocol in sec, calculated from stimuli
-    #    durations.
-#
-    ##    Returns
-    #    -------
-    #    float :
-    #        protocol length in seconds.
-
-    #    """
-    #    total_duration = 0
-    #    for stim in self.stimuli:
-    #        total_duration += stim.duration
-    #    return total_duration
-
     def print(self):
         """Print protocol sequence.
         """
@@ -379,8 +339,8 @@ class Protocol(ParametrizedQt):
             raise ValueError("Protocol does not have a specified name")
         super().__init__(name="stimulus/protocol/" + self.__class__.name)
 
-        self.pre_pause = Param(0., limits=(0., 10000.))
-        self.post_pause = Param(0., limits=(0., 10000.))
+        self.pre_pause = Param(0.0, limits=(0.0, 10000.0))
+        self.post_pause = Param(0.0, limits=(0.0, 10000.0))
         self.n_repeats = Param(1, limits=(1, 10000))
 
     def _get_stimulus_list(self):