+++ stytra/tracking/eyes.py
@@ -16,7 +16,7 @@ class EyeTrackingMethod(ImageToDataNode):
     name = "eyes"
 
     def __init__(self, *args, **kwargs):
-        super().__init__(*args,  name="eyes_tracking", **kwargs)
+        super().__init__(*args, name="eyes_tracking", **kwargs)
 
         headers = []
         for i in range(2):
@@ -35,9 +35,7 @@ class EyeTrackingMethod(ImageToDataNode):
 
         self.data_log_name = "eye_track"
 
-        self.diagnostic_image_options = [
-            "thresholded",
-        ]
+        self.diagnostic_image_options = ["thresholded"]
 
     def _process(
         self,
@@ -68,10 +66,15 @@ class EyeTrackingMethod(ImageToDataNode):
         PAD = 0
 
         cropped = _pad(
-            (im[
-                wnd_pos[1] : wnd_pos[1] + wnd_dim[1],
-                wnd_pos[0] : wnd_pos[0] + wnd_dim[0],
-            ] < threshold).view(dtype=np.uint8).copy(),
+            (
+                im[
+                    wnd_pos[1] : wnd_pos[1] + wnd_dim[1],
+                    wnd_pos[0] : wnd_pos[0] + wnd_dim[0],
+                ]
+                < threshold
+            )
+            .view(dtype=np.uint8)
+            .copy(),
             padding=PAD,
             val=255,
         )
@@ -94,10 +97,7 @@ class EyeTrackingMethod(ImageToDataNode):
                 + e[1][1][::-1]
                 + (-e[1][2],)
             )
-        return NodeOutput(
-            [message, ],
-            self._output_type(*e)
-        )
+        return NodeOutput([message], self._output_type(*e))
 
 
 def _pad(im, padding=0, val=0):